pyproject.toml
[project]
name = "practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "litellm>=1.80.11",
    "openai-agents>=0.2.9",
    "playwright>=1.53.0",
    "python-dotenv>=1.1.1",
]
